refactor: PR 요구 사항 반영
