+++ app/src/main/java/com/example/swapit/data/datasource/RemoteShoppingDataSource.kt
@@ -1,12 +1,11 @@
 package com.example.swapit.data.datasource
 
+import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
+import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingCardProductsResults
 import com.example.swapit.data.datasource.remote.service.ShoppingService
-import com.example.swapit.domain.model.shopping.ShoppingProductResponse
-import com.example.swapit.domain.model.shopping.toDomainModel
 
 class RemoteShoppingDataSource(private val shoppingService: ShoppingService) {
-    suspend fun getRemoteShoppingProductResponse(): ShoppingProductResponse {
-        val response = shoppingService.getGoodsListResponse().toDomainModel()
-        return response
+    suspend fun shoppingProductResponse(): BaseResponse<ShoppingCardProductsResults> {
+        return shoppingService.shoppingProductsResponse()
     }
 }
