--- app/src/main/java/com/example/swapit/data/datasource/remote/dto/response/shopping/GoodsListResponse.kt
@@ -1,10 +1,0 @@
-package com.example.swapit.data.datasource.remote.dto.response.shopping
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GoodsListResponse(
-    val success: Boolean,
-    val message: String,
-    val results: GoodsListResults,
-)
