+++ app/src/main/java/com/example/swapit/data/datasource/remote/dto/response/shopping/ShoppingCardProductsResults.kt
@@ -3,8 +3,8 @@ package com.example.swapit.data.datasource.remote.dto.response.shopping
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GoodsListResults(
-    val goodsList: List<Goods>,
+data class ShoppingCardProductsResults(
+    val goodsList: List<ShoppingCardProduct>,
     val hasNext: Boolean,
     val lastCursorId: Int,
     val size: Int,
