+++ app/src/main/java/com/example/swapit/data/datasource/remote/service/ShoppingService.kt
@@ -1,9 +1,10 @@
 package com.example.swapit.data.datasource.remote.service
 
-import com.example.swapit.data.datasource.remote.dto.response.shopping.GoodsListResponse
+import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
+import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingCardProductsResults
 import retrofit2.http.GET
 
 interface ShoppingService {
     @GET("api/all/goods")
-    suspend fun getGoodsListResponse(): GoodsListResponse
+    suspend fun shoppingProductsResponse(): BaseResponse<ShoppingCardProductsResults>
 }
