+++ app/src/main/java/com/example/swapit/data/repository/DefaultShoppingRepository.kt
@@ -1,7 +1,8 @@
 package com.example.swapit.data.repository
 
 import com.example.swapit.data.datasource.RemoteShoppingDataSource
-import com.example.swapit.domain.model.shopping.ShoppingProductResponse
+import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
+import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingCardProductsResults
 import com.example.swapit.domain.model.shopping.ShoppingProductResults
 import com.example.swapit.domain.model.shopping.toDomainModel
 import com.example.swapit.domain.repository.ShoppingRepository
@@ -10,12 +11,8 @@ class DefaultShoppingRepository(
     private val remoteSource: RemoteShoppingDataSource,
 ) :
     ShoppingRepository {
-    override suspend fun getShoppingResponse(): ShoppingProductResponse {
-        return remoteSource.getRemoteShoppingProductResponse()
+    override suspend fun shoppingCardResults(): ShoppingProductResults {
+        return remoteSource.shoppingProductResponse().results.toDomainModel()
     }
 
-    override suspend fun getShoppingResults(shoppingProductResponse: ShoppingProductResponse): ShoppingProductResults {
-        val results = shoppingProductResponse.results.toDomainModel()
-        return results
-    }
 }
