+++ app/src/main/java/com/example/swapit/domain/model/shopping/ShoppingProductResults.kt
@@ -1,16 +1,16 @@
 package com.example.swapit.domain.model.shopping
 
-import com.example.swapit.data.datasource.remote.dto.response.shopping.Goods
-import com.example.swapit.data.datasource.remote.dto.response.shopping.GoodsListResults
+import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingCardProduct
+import com.example.swapit.data.datasource.remote.dto.response.shopping.ShoppingCardProductsResults
 
 data class ShoppingProductResults(
-    val goodsList: List<Goods>,
+    val goodsList: List<ShoppingCardProduct>,
     val hasNext: Boolean,
     val lastCursorId: Int,
     val count: Int,
 )
 
-fun GoodsListResults.toDomainModel(): ShoppingProductResults {
+fun ShoppingCardProductsResults.toDomainModel(): ShoppingProductResults {
     return ShoppingProductResults(
         goodsList = this.goodsList,
         hasNext = this.hasNext,
