+++ app/src/main/java/com/example/swapit/ui/navigation/NavigationModule.kt
@@ -1,5 +1,6 @@
 package com.example.swapit.ui.navigation
 
+import ShoppingViewModel
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
