+++ app/src/main/java/com/example/swapit/ui/shopping/ShoppingScreen.kt
@@ -34,7 +34,6 @@ fun ShoppingScreen(
     viewModel: ShoppingViewModel,
 ) {
     val sheetState = rememberModalBottomSheetState()
-    viewModel.loadProducts()
 
     Scaffold(
         topBar = {
